+++ CHANGELOG.md
@@ -105,6 +105,10 @@ version 0.2.0, thus covering only a fraction of the project’s history.
     - The ability to target ports, including port holders, and the plate of an
       MCU lock, including a set of vertical segments. The rear housing has also
       gotten vertical segments.
+- The ability to anchor things to a wider array of the parts of a key mount,
+  including different sides of its wall, using cardinal and intercardinal
+  compass points, whereas before, only points intermediate between cardinals
+  and intercardinals could be used.
 - The ability to override specific coordinates for secondary named positions.
 - A `segment` parameter for anchoring wrist-rest mounts.
 - A GNU Make target for the Dactyl-ManuForm.